+++ common/src/main/java/dev/terminalmc/clientsort/client/gui/widget/ControlButton.java
@@ -18,23 +18,25 @@
 package dev.terminalmc.clientsort.client.gui.widget;
 
 import com.mojang.blaze3d.platform.InputConstants;
-import dev.terminalmc.clientsort.client.ClientSort;
+import dev.terminalmc.clientsort.client.config.ButtonLayout;
 import dev.terminalmc.clientsort.client.config.Vec2i;
 import dev.terminalmc.clientsort.client.gui.screen.edit.ContainerPositionEditScreen;
 import dev.terminalmc.clientsort.client.gui.screen.edit.PlayerPositionEditScreen;
 import dev.terminalmc.clientsort.client.gui.screen.edit.PositionEditScreen;
 import dev.terminalmc.clientsort.mixin.client.accessor.AbstractContainerScreenAccessor;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.ComponentPath;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.components.WidgetSprites;
-import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.gui.navigation.FocusNavigationEvent;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.Container;
 import net.minecraft.world.inventory.Slot;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public abstract class ControlButton extends Button {
     public static final int WIDTH = 13;
@@ -42,9 +44,9 @@ public abstract class ControlButton extends Button {
     public static final int HALF_WIDTH = WIDTH / 2;
     public static final int HALF_HEIGHT = HEIGHT / 2;
 
-    private final Screen screen;
+    private final AbstractContainerScreen<?> screen;
     public final Container container;
-    public final String layoutKey;
+    public final @Nullable String layoutKey;
     final boolean isPlayerInv;
     private final Slot referenceSlot;
     private final WidgetSprites sprites;
@@ -54,7 +56,7 @@ public abstract class ControlButton extends Button {
     protected ControlButton(
             AbstractContainerScreen<?> screen,
             Container container,
-            String layoutKey,
+            @Nullable String layoutKey,
             boolean isPlayerInv,
             Slot referenceSlot,
             WidgetSprites sprites,
@@ -63,7 +65,7 @@ public abstract class ControlButton extends Button {
             boolean active
     ) {
         super(
-                ((AbstractContainerScreenAccessor)screen).getLeftPos()
+                ((AbstractContainerScreenAccessor) screen).getLeftPos()
                         + ((AbstractContainerScreenAccessor) screen).getImageWidth()
                         + offset.x(),
                 ((AbstractContainerScreenAccessor) screen).getTopPos()
@@ -88,10 +90,14 @@ public abstract class ControlButton extends Button {
     @Override
     public boolean mouseClicked(double mouseX, double mouseY, int mouseButton) {
         if (isMouseOver(mouseX, mouseY)) {
-            if (mouseButton == InputConstants.MOUSE_BUTTON_RIGHT) {
-                openEditScreen();
+            boolean rightClick = mouseButton == InputConstants.MOUSE_BUTTON_RIGHT;
+            if (Minecraft.getInstance().screen instanceof PositionEditScreen) {
+                if (rightClick) {
+                    active = !active;
+                }
                 return true;
-            } else if (Minecraft.getInstance().screen instanceof PositionEditScreen) {
+            } else if (rightClick) {
+                openEditScreen();
                 return true;
             }
         }
@@ -103,10 +109,10 @@ public abstract class ControlButton extends Button {
      */
     @Override
     public boolean isMouseOver(double mouseX, double mouseY) {
-        return mouseX >= (double)getX()
-                && mouseY >= (double)getY()
-                && mouseX < (double)(getX() + width)
-                && mouseY < (double)(getY() + height);
+        return mouseX >= (double) getX()
+                && mouseY >= (double) getY()
+                && mouseX < (double) (getX() + width)
+                && mouseY < (double) (getY() + height);
     }
 
     public void openEditScreen() {
@@ -118,7 +124,7 @@ public abstract class ControlButton extends Button {
 
     @Override
     public void renderWidget(@NotNull GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
-        AbstractContainerScreenAccessor acs = (AbstractContainerScreenAccessor)screen;
+        AbstractContainerScreenAccessor acs = (AbstractContainerScreenAccessor) screen;
 
         // Keep visible
         int newX = Math.clamp(
@@ -141,13 +147,36 @@ public abstract class ControlButton extends Button {
 
     @Override
     protected void onDrag(double mouseX, double mouseY, double dragX, double dragY) {
-        AbstractContainerScreenAccessor acs = (AbstractContainerScreenAccessor)screen;
-        int newX = Math.clamp((int)mouseX - HALF_WIDTH, 0, screen.width - WIDTH);
-        int newY = Math.clamp((int)mouseY - HALF_HEIGHT, 0, screen.height - HEIGHT);
+        if (Minecraft.getInstance().screen instanceof PositionEditScreen) {
+            AbstractContainerScreenAccessor acs = (AbstractContainerScreenAccessor) screen;
+            int newX = Math.clamp((int) mouseX - HALF_WIDTH, 0, screen.width - WIDTH);
+            int newY = Math.clamp((int) mouseY - HALF_HEIGHT, 0, screen.height - HEIGHT);
 
-        offset = new Vec2i(
-                newX - (acs.getLeftPos() + acs.getImageWidth()),
-                newY - (acs.getTopPos() + referenceSlot.y)
-        );
+            offset = new Vec2i(
+                    newX - (acs.getLeftPos() + acs.getImageWidth()),
+                    newY - (acs.getTopPos() + referenceSlot.y)
+            );
+        } else {
+            super.onDrag(mouseX, mouseY, dragX, dragY);
+        }
     }
+
+    /**
+     * Disables keyboard navigation.
+     */
+    @Override
+    public ComponentPath nextFocusPath(@NotNull FocusNavigationEvent event) {
+        return null;
+    }
+
+    /**
+     * Disables focusing on click.
+     */
+    @Override
+    public void setFocused(boolean focused) {
+        if (!focused)
+            super.setFocused(false);
+    }
+
+    public abstract boolean getLayoutStatus(ButtonLayout layout);
 }